+++ archetypes/api/_APPNAME_/_APPNAME_-api/src/main/java/applica/_APPNAME_/api/controllers/AuthController.java
@@ -45,7 +45,6 @@ public class AuthController extends LocalizedController {
     }
 
     @GetMapping("/freshToken")
-    @PreAuthorize("isAuthenticated()")
     public Response freshToken(HttpServletRequest request) {
         try {
             String currentToken = request.getHeader("TOKEN");