+++ archetypes/client/app/js/api/utils.js
@@ -2,10 +2,20 @@
 
 import * as http from "../aj/http"
 import * as responses from "./responses"
+import {getSessionToken} from "./session"
+import * as _ from "../libs/underscore"
 
-export function post(url, data) {
-    return new Promise((resolve, reject, headers = {}) => {
-        http.post(url, data, headers)
+function addToken(headers) {
+    if (!_.isEmpty(getSessionToken())) {
+        return _.assign(headers || {}, {"x-auth-token": getSessionToken()})
+    } else {
+        return headers
+    }
+}
+
+export function post(url, data, headers = {}) {
+    return new Promise((resolve, reject) => {
+        http.post(url, data, addToken(headers))
             .then(json => {
                 if (_.isEmpty(json)) {
                     reject(responses.ERROR)
@@ -30,7 +40,7 @@ export function postJson(url, data, headers = {}) {
     return new Promise((resolve, reject) => {
         let json = typeof(data) == "string" ? data : JSON.stringify(data)
         headers = _.assign(headers, {"Content-Type": "application/json"})
-        http.post(url, json, headers)
+        http.post(url, json, addToken(headers))
             .then(json => {
                 if (_.isEmpty(json)) {
                     reject(responses.ERROR)
@@ -53,7 +63,7 @@ export function postJson(url, data, headers = {}) {
 
 export function get(url, data, headers = {}) {
     return new Promise((resolve, reject) => {
-        http.get(url, data, headers)
+        http.get(url, data, addToken(headers))
             .then(json => {
                 if (_.isEmpty(json)) {
                     reject(responses.ERROR)
@@ -76,7 +86,7 @@ export function get(url, data, headers = {}) {
 
 export function delete_(url, data, headers) {
     return new Promise((resolve, reject) => {
-        http.delete(url, data, headers)
+        http.delete(url, data, addToken(headers))
             .then(json => {
                 if (_.isEmpty(json)) {
                     reject(responses.ERROR)