+++ archetypes/client/app/js/web/components/loader.jsx
@@ -12,7 +12,7 @@ class PageLoader extends React.Component {
 
     render() {
         return (
-            <div className="page-loader" style={{display: "block"}}>
+            <div className="page-loader" style={{display: "none"}}>
                 <div className="preloader">
                     <svg className="pl-circular" viewBox="25 25 50 50">
                         <circle className="plc-path" cx="50" cy="50" r="20" />