+++ framework-security/src/main/java/applica/framework/security/token/TokenAuthenticationEntryPoint.java
@@ -1,0 +1,22 @@
+package applica.framework.security.token;
+
+import applica.framework.security.Security;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.www.BasicAuthenticationEntryPoint;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+/**
+ * Created by bimbobruno on 27/01/2017.
+ */
+public class TokenAuthenticationEntryPoint extends BasicAuthenticationEntryPoint {
+
+    @Override
+    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
+        response.sendError(HttpServletResponse.SC_UNAUTHORIZED);
+    }
+}