+++ CHANGELOG.md
@@ -31,3 +31,6 @@
 - **Fileserver**
     Fileserver base path can be configured via options.properties
     SimpleFileServer update to avoid deprecated methods
+    
+- **Token login**
+    Filter changed to attach in spring security filter chain